- fixed warnings

+++ hi_streaming/hi_streaming/MonolithAudioFormat.h
@@ -399,7 +399,7 @@ public:
 
 	int64 getMonolithLength(int sampleIndex) const
 	{
-		return jmax<int>(0, multiChannelSampleInformation[0][sampleIndex].length);
+		return jmax<int64>(0, multiChannelSampleInformation[0][sampleIndex].length);
 	}
 
 	double getMonolithSampleRate(int sampleIndex) const
